+++ internal/utils/checks_test.go
@@ -141,7 +141,7 @@ func TestPreCheck(t *testing.T) {
 				return tc.runner
 			})
 
-			test.PrepareTestCaseDir(filepath.Join("testdata", "checks_tests", tc.caseTestDataDir, "docker-compose-scripts"), tc.path)
+			test.PrepareTestCaseDir(filepath.Join("testdata", "checks_tests", tc.caseTestDataDir, configs.DefaultSedgeDataFolderName), tc.path)
 
 			err := PreCheck(tc.path)
 			if tc.isErr && err == nil {
@@ -169,7 +169,7 @@ type checkContainersTC struct {
 
 func buildCheckContainersTestCase(t *testing.T, caseName string, isErr bool) *checkContainersTC {
 	dcPath := t.TempDir()
-	test.PrepareTestCaseDir(filepath.Join("testdata", "checks_tests", caseName, "docker-compose-scripts"), dcPath)
+	test.PrepareTestCaseDir(filepath.Join("testdata", "checks_tests", caseName, configs.DefaultSedgeDataFolderName), dcPath)
 
 	tc := checkContainersTC{}
 	tc.path = dcPath